backend/requirements.txt
Flask==2.0.1 
flask-cors==3.0.10 
sentence-transformers==2.7.0 
typesense==0.12.0 
python-dotenv==0.19.0 
numpy==1.21.0 
scipy==1.7.0 
torch==2.0.1 
werkzeug==2.0.3 
huggingface_hub==0.23.0